Stable version (needs more documentation)
New quality control system in the the list of peaks position.
Reason: The software needs some properties of the peaks like indexes of
both corners (index where the signal reaches the signal mean).  In case
the peaks are too close from the border, it is impossible to get both
sides indexes.

The output file is now in CSV format. First item is the time
YYYYmmDDHHMMss when the event happens and the second index is the SNR
defined as (peak maximum - mean)/mean.

Future activities:
Better internal documentation
Create a new file with parameters as headers